# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PageFire API
  version: 0.3.0
  description: |
    PageFire is a unified monitoring and incident management platform.
    Single Go binary, self-hosted, MIT license.

    Authentication is via cookie-based sessions (login endpoint) or
    Bearer API tokens (created via the tokens endpoint). Admin-only
    write operations are enforced on users, teams, services, escalation
    policies, and schedules.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    url: https://pagefire.dev
servers:
  - url: /api/v1
    description: Default API base path
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /alerts:
    post:
      operationId: createAlert
      tags: [Alerts]
      summary: Create a new alert
      description: Triggers the escalation pipeline for the specified service.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertRequest"
      responses:
        "201":
          description: Alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listAlerts
      tags: [Alerts]
      summary: List alerts with filtering and pagination
      parameters:
        - name: status
          in: query
          description: "Filter by status: triggered, acknowledged, resolved"
          schema:
            type: string
            enum: [triggered, acknowledged, resolved]
        - name: service_id
          in: query
          description: Filter by service ID
          schema:
            type: string
        - name: group_key
          in: query
          description: Filter by group key
          schema:
            type: string
        - name: source
          in: query
          description: "Filter by source: api, integration, monitor, test"
          schema:
            type: string
        - name: search
          in: query
          description: Full-text search in summary (max 200 chars)
          schema:
            type: string
            maxLength: 200
        - name: created_after
          in: query
          description: Filter alerts created after this time (RFC3339)
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          description: Filter alerts created before this time (RFC3339)
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Maximum number of results (default 50, max 200)
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          description: Offset for pagination
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    get:
      operationId: getAlert
      tags: [Alerts]
      summary: Get an alert by ID
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Alert found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          $ref: "#/components/responses/NotFound"
  /alerts/{id}/acknowledge:
    post:
      operationId: acknowledgeAlert
      tags: [Alerts]
      summary: Acknowledge an alert
      description: Stops escalation. user_id is optional — defaults to the authenticated user.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  description: User acknowledging (defaults to session user)
      responses:
        "200":
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /alerts/{id}/resolve:
    post:
      operationId: resolveAlert
      tags: [Alerts]
      summary: Resolve an alert
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Alert resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /alerts/{id}/logs:
    get:
      operationId: listAlertLogs
      tags: [Alerts]
      summary: List audit log entries for an alert
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Alert log list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertLog"
  /alerts/{id}/incident:
    get:
      operationId: getIncidentForAlert
      tags: [Alerts]
      summary: Get the incident linked to an alert
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Linked incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          $ref: "#/components/responses/NotFound"
        "501":
          description: Incident linking not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ──────────────────────────── Incidents ────────────────────────────
...
  /incidents/{id}/alerts:
    get:
      operationId: listIncidentAlerts
      tags: [Incidents]
      summary: List alerts linked to an incident
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Linked alert list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: linkAlertToIncident
      tags: [Incidents]
      summary: Link an alert to an incident
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [alert_id]
              properties:
                alert_id:
                  type: string
      responses:
        "204":
          description: Alert linked
        "400":
          $ref: "#/components/responses/BadRequest"
  /incidents/{id}/alerts/{alertID}:
    delete:
      operationId: unlinkAlertFromIncident
      tags: [Incidents]
      summary: Unlink an alert from an incident
      parameters:
        - $ref: "#/components/parameters/id"
        - name: alertID
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Alert unlinked
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────── On-Call ────────────────────────────
...
  /integrations/{key}/alerts:
    post:
      operationId: genericWebhook
      tags: [Integrations]
      summary: Generic inbound webhook
      description: |
        Accepts alert payloads and creates alerts for the service associated with
        the integration key. Authenticated by the key secret in the URL path.
        Rate-limited to 60 requests per minute.
      security: []
      parameters:
        - name: key
          in: path
          required: true
          description: Integration key secret
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericWebhookRequest"
      responses:
        "201":
          description: Alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "200":
          description: Deduplicated — existing alert returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid integration key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: pagefire_session
      description: Server-side session cookie set by POST /auth/login
    bearerAuth:
      type: http
      scheme: bearer
      description: API token created via POST /auth/tokens
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    StatusResponse:
      type: object
      properties:
        status:
          type: string
      required: [status]

    # ── Users ──
...
    Alert:
      type: object
      properties:
        id:
          type: string
        service_id:
          type: string
        status:
          type: string
          enum: [triggered, acknowledged, resolved]
        summary:
          type: string
          maxLength: 500
        details:
          type: string
          maxLength: 10000
        source:
          type: string
          enum: [api, integration, monitor, test]
        dedup_key:
          type: string
        group_key:
          type: string
        escalation_policy_snapshot:
          type: string
          description: JSON-encoded snapshot of the escalation policy at alert creation time
        escalation_step:
          type: integer
        loop_count:
          type: integer
        next_escalation_at:
          type: string
          format: date-time
          nullable: true
        acknowledged_by:
          type: string
          nullable: true
        acknowledged_at:
          type: string
          format: date-time
          nullable: true
        resolved_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, service_id, status, summary, source, escalation_step, loop_count, created_at]
    CreateAlertRequest:
      type: object
      required: [service_id, summary]
      properties:
        service_id:
          type: string
        summary:
          type: string
          maxLength: 500
        details:
          type: string
          maxLength: 10000
        dedup_key:
          type: string
        group_key:
          type: string
    AlertLog:
      type: object
      properties:
        id:
          type: string
        alert_id:
          type: string
        event:
          type: string
          enum: [created, escalated, acknowledged, resolved]
        message:
          type: string
        user_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, alert_id, event, message, created_at]

    # ── Incidents ──
    Incident:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
          maxLength: 500
        status:
          type: string
          enum: [triggered, acknowledged, investigating, identified, monitoring, resolved]
        severity:
          type: string
          enum: [critical, major, minor]
        summary:
          type: string
          maxLength: 10000
        source:
          type: string
          enum: [manual, alert, monitor]
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
          nullable: true
      required: [id, title, status, severity, source, created_at]
...
    GenericWebhookRequest:
      type: object
      required: [summary]
      properties:
        summary:
          type: string
          maxLength: 500
        details:
          type: string
          maxLength: 10000
        dedup_key:
          type: string
        group_key:
          type: string
...